Implement 'batteryLevel', 'batteryState', and 'orientation'. This is broken at the moment.. kinda.. only the first call seems to work. It seems like after that, the values are cached...
I believe this is due to the fact that the [NSRunLoop mainRunLoop] isn't running in your node process, as I have written a simple ObjC 'hello world' run loop
program that prints out these properties on a repeating NSTimer and the values *do* change. So it looks like until node (i.e. libev) has an NSRunLoop backend,
we're SOL.

On that note, these three properties will remain undocumented for now