Stop node from crashing on 'getContacts()'.
It seems like a double release was happening. But I have no idea what the fuck I'm
with all these malloc() and free() calls, so if anybody good with C memory
management wants to look over this file, then please do!

I think this leaks memory now, but at least we're not crashing...