More progress on the AddressBook stuff.
Now the 'firstName' and 'lastName' properties of Contact instances are properly
retrieved from the address book database. No crashing. Maybe I'm starting to
get the hang of this stuff :)

The only thing I don't like is that I had to pussy out and use NSString* instead
of CFStringRef, so that required an NSAutoreleasePool* to be created on the
background thread. Not too big of a deal, but maybe figure out the low-level stuff
someday...